requirements.txt
pygithub
daiquiri
requests
